+++ base/dists/gumbel/examples/index.js
@@ -21,7 +21,7 @@
 var Float64Array = require( '@stdlib/array/float64' );
 var filledarrayBy = require( '@stdlib/array/filled-by' );
 var mean = require( './../../../../strided/mean' );
-var variance = require( './../../../../base/variance' );
+var variance = require( './../../../../strided/variance' );
 var stdev = require( './../../../../base/stdev' );
 var randGumbel = require( '@stdlib/random/base/gumbel' ).factory;
 var gumbel = require( './../lib' );